fix: update code and storybook based on pr feedback
