+++ src/__stories__/RealWorld.tsx
@@ -8,7 +8,6 @@ import { JsonSchemaProps, JsonSchemaViewer } from '../components/JsonSchemaViewe
 const defaultSchema = require('../__fixtures__/default-schema.json');
 const boxFileSchema = require('../__fixtures__/real-world/box-file.json');
 const githubIssueSchema = require('../__fixtures__/real-world/github-issue.json');
-const extensionsSchema = require('../__fixtures__/real-world/vendor-extensions.json');
 
 export default {
   component: JsonSchemaViewer,
@@ -27,18 +26,4 @@ GithubIssueSchema.args = {
   schema: githubIssueSchema as JSONSchema4,
   parentCrumbs: ['Github', 'Issue'],
   renderRootTreeLines: true,
-};
-
-export const ExtensionRowSchema = Template.bind({});
-ExtensionRowSchema.args = {
-  schema: extensionsSchema as JSONSchema4,
-  defaultExpandedDepth: Infinity,
-  renderRootTreeLines: true,
-  renderExtensionAddon: () => {
-    return (
-      <Flex h="full" alignItems="center">
-        Vendor Extensions
-      </Flex>
-    );
-  },
 };
